backend/functions/nasa_photo_fetcher/requirements.txt
boto3==1.37.38
requests==2.32.3
psycopg==3.2.4
psycopg-binary==3.2.4
python-dotenv==1.0.1
